fix(naas/runner/jobs.py): Fix job not returning "runs" because of a JSON exception being triggered
Added a type verification on "runs" property of a Job when listing them. It seems that most of the time they are already loaded into Python objects and we were forcing the use of `json.loads()` on them which was raising an exception. We are now type checking to see if we need to run json.loads or not. We are also printing the Exception on stdout if we catch any.